Move partial trait impls to `partial::objects`
